server/requirements.txt
flask
pandas
openpyxl
python-dateutil
psycopg2-binary 
flask-cors
python-dotenv
gunicorn
jinja2